Actualización completa: controladores, modelos, migraciones, vistas, pruebas unitarias y funcionales

+++ database/migrations/2025_07_23_031101_add_usuario_responsable_to_metas_table.php
@@ -6,16 +6,11 @@ use Illuminate\Support\Facades\Schema;
 
 class AddUsuarioResponsableToMetasTable extends Migration
 {
-    /**
-     * Ejecutar las migraciones.
-     */
     public function up()
     {
         Schema::table('metas', function (Blueprint $table) {
-            // Cambiado a integer para que coincida con 'IdUsuario' que es int
-            $table->integer('usuario_responsable_id')->nullable()->after('estado');
+            $table->unsignedBigInteger('usuario_responsable_id')->nullable()->after('estado');
 
-            // Clave foránea hacia la tabla usuarios con la columna correcta
             $table->foreign('usuario_responsable_id')
                   ->references('IdUsuario')
                   ->on('usuarios')
@@ -23,9 +18,6 @@ class AddUsuarioResponsableToMetasTable extends Migration
         });
     }
 
-    /**
-     * Revertir las migraciones.
-     */
     public function down()
     {
         Schema::table('metas', function (Blueprint $table) {
